+++ CRM/DataprocessorSearch/Form/ContributionSearch.php
@@ -1,0 +1,110 @@
+<?php
+/**
+ * @author Jaap Jansma <jaap.jansma@civicoop.org>
+ * @license AGPL-3.0
+ */
+
+use CRM_Dataprocessor_ExtensionUtil as E;
+
+class CRM_DataprocessorSearch_Form_ContributionSearch extends CRM_DataprocessorSearch_Form_AbstractSearch {
+
+  /**
+   * Returns the name of the default Entity
+   *
+   * @return string
+   */
+  public function getDefaultEntity() {
+    return 'Contribution';
+  }
+
+  /**
+   * Returns the url for view of the record action
+   *
+   * @param $row
+   *
+   * @return false|string
+   */
+  protected function link($row) {
+    return CRM_Utils_System::url('civicrm/contact/view/contribution', 'reset=1&id='.$row['id'].'&cid='.$row['id'].'&action=view');
+  }
+
+  /**
+   * Returns the link text for view of the record action
+   *
+   * @param $row
+   *
+   * @return false|string
+   */
+  protected function linkText($row) {
+    return E::ts('View contribution');
+  }
+
+  /**
+   * Checks whether the output has a valid configuration
+   *
+   * @return bool
+   */
+  protected function isConfigurationValid() {
+    if (!isset($this->dataProcessorOutput['configuration']['contribution_id_field'])) {
+      return false;
+    }
+    return true;
+  }
+
+  /**
+   * Return the data processor ID
+   *
+   * @return String
+   */
+  protected function getDataProcessorName() {
+    $dataProcessorName = str_replace('civicrm/dataprocessor_contribution_search/', '', CRM_Utils_System::getUrlPath());
+    return $dataProcessorName;
+  }
+
+  /**
+   * Returns the name of the output for this search
+   *
+   * @return string
+   */
+  protected function getOutputName() {
+    return 'contribution_search';
+  }
+
+  /**
+   * Returns the name of the ID field in the dataset.
+   *
+   * @return string
+   */
+  protected function getIdFieldName() {
+    return $this->dataProcessorOutput['configuration']['contribution_id_field'];
+  }
+
+  /**
+   * @return string
+   */
+  protected function getEntityTable() {
+    return 'civicrm_contribution';
+  }
+
+  /**
+   * Returns whether we want to use the prevnext cache.
+   * @return bool
+   */
+  protected function usePrevNextCache() {
+    return true;
+  }
+
+  /**
+   * Builds the list of tasks or actions that a searcher can perform on a result set.
+   *
+   * @return array
+   */
+  public function buildTaskList() {
+    if (!$this->_taskList) {
+      $taskParams['softCreditFiltering'] = FALSE;
+      $this->_taskList = CRM_Contribute_Task::permissionedTaskTitles(CRM_Core_Permission::getPermission(), $taskParams);
+    }
+    return $this->_taskList;
+  }
+
+}